Fixing stuff for mne bids.
