+++ doc/api.rst
@@ -37,7 +37,9 @@ MNE BIDS
    get_head_mri_trans
    get_matched_empty_room
    get_anonymization_daysback
-
+   print_dir_tree
+   get_entity_vals
+   get_kinds
 
 Path
 ====
@@ -53,11 +55,7 @@ Path
 .. autosummary::
    :toctree: generated/
 
-   print_dir_tree
-   get_entity_vals
-   get_kinds
    BIDSPath
-
 
 Copyfiles
 =========
